machine_learning/labs/requirements.txt
absl-py==1.0.0
autopep8==1.6.0
black==22.3.0
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.3
google-auth==2.6.6
google-auth-oauthlib==0.4.6
grpcio==1.46.1
idna==3.3
importlib-metadata==4.11.3
joblib==1.1.0
Markdown==3.3.7
mypy-extensions==0.4.3
numpy==1.21.6
oauthlib==3.2.0
pandas==1.3.5
pathspec==0.9.0
Pillow==9.1.0
platformdirs==2.5.2
protobuf==3.20.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
scikit-learn==1.0.2
scipy==1.7.3
six==1.16.0
tensorboard==2.9.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
threadpoolctl==3.1.0
toml==0.10.2
tomli==2.0.1
torch==1.8.0
torch-tb-profiler==0.4.0
torchvision==0.9.0
tqdm==4.64.0
typed-ast==1.5.3
typing_extensions==4.2.0
urllib3==1.26.9
Werkzeug==2.1.2
yapf==0.32.0
zipp==3.8.0
